+++ dev_utils/mockServer.ts
@@ -22,12 +22,22 @@ export function startMockServer2(): {
   mqttConn2: MqttConn;
   mqttServer: MqttServer;
 } {
-  const mqttServer = new MqttServer({ handlers });
-  const { input: input1, output: output1 } = createWebStreamPair();
+  const { mqttConn: mqttConn1, mqttServer } = startMockServer();
   const { input: input2, output: output2 } = createWebStreamPair();
-  const mqttConn1 = new MqttConn({ conn: output1 });
   const mqttConn2 = new MqttConn({ conn: output2 });
-  mqttServer.serve(input1);
   mqttServer.serve(input2);
   return { mqttConn1, mqttConn2, mqttServer };
+}
+
+export function startMockServer3(): {
+  mqttConn1: MqttConn;
+  mqttConn2: MqttConn;
+  mqttConn3: MqttConn;
+  mqttServer: MqttServer;
+} {
+  const { mqttConn1, mqttConn2, mqttServer } = startMockServer2();
+  const { input: input3, output: output3 } = createWebStreamPair();
+  const mqttConn3 = new MqttConn({ conn: output3 });
+  mqttServer.serve(input3);
+  return { mqttConn1, mqttConn2, mqttConn3, mqttServer };
 }