+++ dev_utils/mod.ts
@@ -7,13 +7,19 @@ export {
   runAsap,
   withTimeout,
 } from "./timers.ts";
-export { startMockServer, startMockServer2 } from "./mockServer.ts";
+export {
+  startMockServer,
+  startMockServer2,
+  startMockServer3,
+} from "./mockServer.ts";
 export {
   checkNoPacket,
   connect,
   disconnect,
   nextPacketWithTimeOut,
   ping,
+  publish,
   subscribe,
 } from "./packetHelpers.ts";
 export { generateLocalhostCerts } from "./generateCert.ts";
+export { handlers, isAuthenticatedBroker } from "./test-handlers.ts";